+++ src/backend/main/java/com/retrobased/market/repositories/RepositoryCliente.java
@@ -11,7 +11,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface RepositoryCliente extends JpaRepository<Cliente,Integer> {
+public interface RepositoryCliente extends JpaRepository<Cliente,Long> {
     Cliente findByEmail(String email);
     boolean existsByEmail(String email);
 
@@ -19,13 +19,13 @@ public interface RepositoryCliente extends JpaRepository<Cliente,Integer> {
             "FROM Ordine o " +
             "WHERE o.idCliente = ?1 "
     )
-    List<Ordine> findOrdiniByIdCliente(Integer idCliente);
+    List<Ordine> findOrdiniByIdCliente(Long idCliente);
 
     @Query("SELECT po.idProdotto " +
             "FROM ProdottoOrdinato po JOIN po.idOrdine o " +
             "WHERE o.idCliente = ?1"
     )
-    List<Prodotto> findOrderedItemsByUser(Integer id); // TODO CAMBIARE CON TOKEN
+    List<Prodotto> findOrderedItemsByUser(Long id); // TODO CAMBIARE CON TOKEN
 
     // List<Prodotto> findProdottoByIdCliente(Integer idCliente);
