+++ src/backend/main/java/com/retrobased/market/repositories/RepositoryIndirizzoCliente.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface RepositoryIndirizzoCliente extends JpaRepository<IndirizzoCliente,Integer> {
 
-    boolean existsIndirizzoClienteByIdAndIdCliente(Integer idIndirizzo, Cliente idCliente);
+    boolean existsIndirizzoClienteByIdAndIdCliente(Long idIndirizzo, Cliente idCliente);
 }
